FSDS with Gen AI Assignment-4/JARVIS-Voice-Assistant/requirements.txt
SpeechRecognition==3.14.3 
pyttsx3==2.90
pyaudio
wikipedia
google-generativeai
pyautogui